试验检测-处理

+++ src/main/java/com/lims/manage/erp/controller/TaskController.java
@@ -385,6 +385,12 @@ public class TaskController {
             paramVo.setReportProducer(userInfo.getName());
             //领样人
             paramVo.setSampler(userInfo.getName());
+            // 见习生：实习的新手
+            paramVo.setProbationer(userInfo.getName());
+            // 实习生
+            paramVo.setInterns(userInfo.getName());
+            // 辅助人员
+            paramVo.setAuxiliaryPersonnel(userInfo.getName());
             return ResultUtil.success("查询任务列表成功！", taskService.getTaskListTwo(paramVo, deptIds));
         }
     }
@@ -965,38 +971,47 @@ public class TaskController {
     @RequestMapping(value = "/checkItemReview")
     public void previewDownLoad(String list , HttpServletResponse response) throws Exception {
         String newFilePath = qiYueSuoEntity.getAutographPath() + GenID.getID() + ".xlsx";
-        String path = qiYueSuoEntity.getAutographPath()+GenID.getID()+".pdf";
+        String path = qiYueSuoEntity.getAutographPath() + GenID.getID() + ".pdf";
         ExcelInsertVo excelInsertVo = new ExcelInsertVo();
         String[] items = list.split(",");
-        List<Integer> idList = new ArrayList<>();
+        Integer[] ids = new Integer[items.length];
         for (int j = 0; j < items.length; j++) {
-            idList.add(Integer.parseInt(items[j]));
+            ids[j] = Integer.parseInt(items[j]);
         }
-        excelInsertVo.setList(idList);
-        Integer[] ids =new Integer[excelInsertVo.getList().size()];
         // 查询检测项对应的 sheet下标
         List<ExcelInsertVo> sheetItems = testProductItemDao.selectItemSheetIndex(ids);
-        if(!CollectionUtils.isEmpty(sheetItems)){
-            // excel 转 pdf
-            XSSFWorkbook wb = taskService.getOriginalRecordAttachment(excelInsertVo);
-            FileOutputStream out = new FileOutputStream(newFilePath);
-            wb.write(out);
-            out.flush();//刷新
-            InputStream out000 = new FileInputStream(newFilePath);
-            //相应pdf
-            ByteArrayOutputStream b1 = PDFHelper3.excel2pdf(out000,path);
-            InputStream inputStream = FileAndFolderUtil.parseOut(b1);
-            ServletOutputStream outputStream = response.getOutputStream();
-            int i = IOUtils.copy(inputStream, outputStream);   // copy流数据,i为字节数
-            inputStream.close();
-            outputStream.close();
-            out000.close();
-            b1.close();
-            out.close();//关闭
-            // 删除附件
-            FileAndFolderUtil.delete(newFilePath);
-            FileAndFolderUtil.delete(path);
-        }else {
+        if (!CollectionUtils.isEmpty(sheetItems)) {
+            List<Integer> idList = new ArrayList<>();
+            for (int i = 0; i < sheetItems.size(); i++) {
+                ExcelInsertVo excelInsertVo1 = sheetItems.get(i);
+                if (StringUtils.isNotEmpty(excelInsertVo1.getTestSetUrl()) && StringUtils.isNotEmpty(excelInsertVo1.getRecordSetUrl())) {
+                    idList.add(excelInsertVo1.getItemId());
+                }
+            }
+            excelInsertVo.setList(idList);
+            // 判断检测数据不为空
+            if (CollectionUtil.isNotEmpty(excelInsertVo.getList())) {
+                // excel 转 pdf
+                XSSFWorkbook wb = taskService.getOriginalRecordAttachment(excelInsertVo);
+                FileOutputStream out = new FileOutputStream(newFilePath);
+                wb.write(out);
+                out.flush();//刷新
+                InputStream out000 = new FileInputStream(newFilePath);
+                //相应pdf
+                ByteArrayOutputStream b1 = PDFHelper3.excel2pdf(out000, path);
+                InputStream inputStream = FileAndFolderUtil.parseOut(b1);
+                ServletOutputStream outputStream = response.getOutputStream();
+                int i = IOUtils.copy(inputStream, outputStream);   // copy流数据,i为字节数
+                inputStream.close();
+                outputStream.close();
+                out000.close();
+                b1.close();
+                out.close();//关闭
+                // 删除附件
+                FileAndFolderUtil.delete(newFilePath);
+                FileAndFolderUtil.delete(path);
+            }
+        } else {
             // 检测项无Sheet页
 
         }
